Działa wszystko oprócz check

+++ src/main/java/com/example/pythonapp/repository/TeacherRepository.java
@@ -6,8 +6,8 @@ import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 @Repository
-public interface TeacherRepository extends JpaRepository<Teacher,Long> {
+public interface TeacherRepository extends JpaRepository<Teacher,Integer> {
     Teacher findByEmail(String email);
     Optional<Teacher> findByName(String name);
-    Teacher findById(long id);
+    Teacher findById(int id);
 }
