+++ src/main/java/com/example/pythonapp/repository/UserRepository.java
@@ -13,11 +13,11 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<UserEntity,Long> {
+public interface UserRepository extends JpaRepository<UserEntity,Integer> {
     Optional<UserEntity> findByEmail(String email);
     Optional<UserEntity> findByName(String name);
     Optional<UserEntity> findByNameAndPassword(String name, String password);
-    Optional<UserEntity> deleteById(long id);
+    Optional<UserEntity> deleteById(int id);
     @Transactional
     @Modifying
     @Query(value = "update users set password=:password where email=:email", nativeQuery = true)
